requirements.txt
streamlit>=1.35.0
plotly>=5.0.0
pandas>=2.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
python-dotenv>=1.0.0
databricks-sql-connector>=3.0.0
azure-eventhub>=5.11.0
azure-mgmt-eventhub>=10.1.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
faker>=20.0.0
requests>=2.31.0